% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_data_by_design.R
\name{arrange_data_by_design}
\alias{arrange_data_by_design}
\title{Arrange columns of \code{data} into groups}
\usage{
arrange_data_by_design(data, column_design)
}
\arguments{
\item{data}{data to rearrange}

\item{column_design}{a vector indicating group membership of columns}
}
\value{
a list with the data and column design rearranged into groups
}
\description{
Rearranges the columns of a matrix based on \code{column_design}
so that the groups of columns are together
}
\author{
Luke Moraglia
}
