% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_plsmfa.R
\name{boot_plsmfa}
\alias{boot_plsmfa}
\title{Bootstrap for loadings of PLSMFA}
\usage{
boot_plsmfa(
  data1,
  data2,
  column_design1,
  column_design2,
  center1 = TRUE,
  center2 = TRUE,
  scale1 = "SS1",
  scale2 = "SS1",
  n_iter = 1000,
  n_dimensions = 0,
  boot_ratio_threshold = 2
)
}
\arguments{
\item{data1}{The first data matrix, also called "X", with \eqn{N} rows and \eqn{I} columns
organized into \eqn{K} subtables.}

\item{data2}{The second data matrix, also called "Y", with \eqn{N} rows and \eqn{J} columns
organized into \eqn{M} subtables.}

\item{column_design1}{A vector of length \eqn{I} that gives the subtable membership of the X variables}

\item{column_design2}{A vector of length \eqn{J} that gives the subtable membership of the Y variables}

\item{center1}{(default TRUE) Whether to center \code{data1}}

\item{center2}{(default TRUE) Whether to center \code{data2}}

\item{scale1}{(default "SS1") Whether to scale \code{data1}.
Any acceptable input to \code{\link[ExPosition]{expo.scale}}}

\item{scale2}{(default "SS1") Whether to scale \code{data2}
Any acceptable input to \code{\link[ExPosition]{expo.scale}}}

\item{n_iter}{(default 1000) The number of iterations for the bootstrap}

\item{n_dimensions}{(default 0) The number of dimensions to bootstrap. The default 0 means to use all
of the dimensions. For data with many variables, using all of the dimensions will take longer to compute.}

\item{boot_ratio_threshold}{(default 2) The threshold for deeming bootstrap ratios as "significant"}
}
\value{
A list that contains:
  \item{\code{u_boot_cube}}{A 3D array containing \code{n_iter} bootstrapped versions of \code{u}}
  \item{\code{boot_ratios_i}}{Bootstrap ratios for the \eqn{I} variables of \code{data1}}
  \item{\code{sig_boot_ratios_i}}{Whether each bootstrap ratio for the \eqn{I} variables of \code{data1}
   is above the threshold.}
  \item{\code{v_boot_cube}}{A 3D array containing n_iter bootstrapped versions of \code{v}}
  \item{\code{boot_ratios_j}}{Bootstrap ratios for the J variables of \code{data2}}
  \item{\code{sig_boot_ratios_j}}{Whether each bootstrap ratio for the J variables of \code{data2}}
}
\description{
Bootstrap methods for PLSMFA are a beta feature. They have not been systematically validated.
}
\details{
\code{boot_plsmfa} performs bootstrap resampling to generate a sampling distribution of the loadings.
This distribution can be used to generate confidence intervals, or bootstrap ratios.

Bootstrap resampling is used to assess the stability of the results by simulating how the results would
vary if the data were collected again from the same population as the sample data. It simulates this process
by sampling the observations of the data \emph{with replacement}. This "bootstrapped sample" is analyzed
with PLSMFA and the results of the loadings are recorded. \code{boot_plsmfa} repeats these steps
\code{n_iter} times and stores the "cube" of loadings. From this cube, confidence intervals could be
calculated, but more
often (especially when there is a large number of variables), a single value called a bootstrap ratio is
used to assess the stability of the loading. The bootstrap ratio is similar to a $t$-value, since it is
the mean of the sampling distribution divided by the standard deviation of the distribution.

At present, \code{boot_plsmfa} computes bootstrap ratios, but not confidence intervals. Also, as a technical
aside, because the SVD is prone to nuisance variance from reflection, reordering, and rotation (see Milan
and Whittaker, 1995), \code{boot_plsmfa} implements a Procrustes rotation on the loadings of each iteration
(\code{uv_boot}).
}
\author{
Luke Moraglia
}
