% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsmfa_graphs.R
\name{boot_ratio_barplot}
\alias{boot_ratio_barplot}
\title{Create a bar plot of bootstrap ratios (or any numerical vector)}
\usage{
boot_ratio_barplot(
  boot_ratios,
  dimension = 1,
  threshold = 2,
  above_threshold = FALSE,
  color_bar = NULL,
  ...
)
}
\arguments{
\item{boot_ratios}{matrix of data to plot}

\item{dimension}{(default 1) which column to plot}

\item{threshold}{(default 2) bars below threshold are gray}

\item{above_threshold}{(default FALSE) whether to plot only the values above \code{threshold}}

\item{color_bar}{(default NULL) a vector of colors for the bars, the same length as the data}

\item{...}{args passed to \code{\link[PTCA4CATA]{PrettyBarPlot2}}}
}
\value{
The bar plot, output from \code{\link[PTCA4CATA]{PrettyBarPlot2}}
}
\description{
Create a bar plot of bootstrap ratios (or any numerical vector)
}
\author{
Luke Moraglia
}
