% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{contributions}
\alias{contributions}
\title{Compute contributions for variables and subtables}
\usage{
contributions(left_load, right_load, column_design_X, column_design_Y)
}
\arguments{
\item{left_load}{left singular vectors / loadings for X}

\item{right_load}{right singular vectors / loadings for Y}

\item{column_design_X}{design vector for columns of X}

\item{column_design_Y}{design vector for columns of Y}
}
\value{
A list containing:
 \item{\code{ctr_var_X}}{Contributions for variables of X}
 \item{\code{ctr_var_Y}}{Contributions for variables of Y}
 \item{\code{ctr_sub_X}}{Contributions for subtables of X}
 \item{\code{ctr_sub_Y}}{Contributions for subtables of Y}
}
\description{
The function uses the left singular vectors (loadings for X), the right singular
vectors (loadings for Y), and the column design vectors to compute the
contributions for each variable and each subtable.
}
\author{
Luke Moraglia
}
