% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsmfa.R
\name{first_pos_gplssvd}
\alias{first_pos_gplssvd}
\title{Make the first element of each singular vector positive for results of \code{GSVD::gplssvd}}
\usage{
first_pos_gplssvd(res_gplssvd, X, Y)
}
\arguments{
\item{res_gplssvd}{Output from \code{\link[GSVD]{gplssvd}}}

\item{X}{Input to \code{gplssvd}}

\item{Y}{Input to \code{gplssvd}}
}
\value{
Results from gplssvd changed accordingly.
}
\description{
This is a helper function that prevents problems related to axis flipping by setting the first
element of each singular vector in \code{u} as positive.
}
\details{
Each column of \code{u} is set so that the first element is positive. The rest of the results
are adapted accordingly.
}
\author{
Luke Moraglia
}
