% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_latent_variable.R
\name{get_partial_lv_group_means}
\alias{get_partial_lv_group_means}
\title{Compute the group means of the partial latent variables, based on a design vector}
\usage{
get_partial_lv_group_means(partial_lv, design)
}
\arguments{
\item{partial_lv}{A block of partial latent variables. Should be N x dims x subtables}

\item{design}{A design vector for the N observations}
}
\value{
A block of plv means. Groups x dims x subtables
}
\description{
Compute the group means of the partial latent variables, based on a design vector
}
\author{
Luke Moraglia
}
