% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsmfa_graphs.R
\name{latent_variable_map}
\alias{latent_variable_map}
\title{Create a factor map of two latent variables from the same matrix}
\usage{
latent_variable_map(
  res_plsmfa,
  axis1 = 1,
  axis2 = 2,
  XorY = "X",
  design = NULL,
  col_obs = "blueviolet",
  col_group = "blueviolet",
  n_iter = 100,
  xlab = paste("Latent Variable", axis1),
  ylab = paste("Latent Variable", axis2),
  ...
)
}
\arguments{
\item{res_plsmfa}{output of \code{\link{plsmfa}}}

\item{axis1}{(default 1) the latent variable to plot on the horizontal axis}

\item{axis2}{(default 2) the latent variable to plot on the vertical axis}

\item{XorY}{(default "X") Either "X" or "Y", showing which matrix the LVs should come from}

\item{design}{(default NULL) A design vector for the rows. If included, group means are plotted}

\item{col_obs}{(default "blueviolet") Either one color or a vector of colors with a length of N observations}

\item{col_group}{(default "blueviolet") Either one color or a vector of colors with a length of N groups}

\item{n_iter}{(default 100) Number of bootstrap iterations for computing confidence intervals around the group means}

\item{xlab}{(default \code{paste("Latent Variable", axis1)}) Horizontal axis label}

\item{ylab}{(default \code{paste("Latent Variable", axis2)}) Vertical axis label}

\item{...}{args passed to \code{\link[PTCA4CATA]{createFactorMap}} for the lv_fact_map}
}
\value{
A list containing:
  \item{\code{plot}}{The final plot from the function}
  \item{\code{lv_fact_map}}{Output from \code{\link[PTCA4CATA]{createFactorMap}} for the observations}
  \item{\code{labels}}{The axis labels}
  \item{\code{group_means}}{The group means on the latent variables}
  \item{\code{lv_group_means_fact_map}}{Output from \code{\link[PTCA4CATA]{createFactorMap}} for the group means}
  \item{\code{CI_ellipses}}{The confidence intervals of the group means}
}
\description{
Create a factor map of two latent variables from the same matrix
}
\author{
Luke Moraglia
}
