% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsmfa_graphs.R
\name{loading_plot}
\alias{loading_plot}
\title{Create plot of loadings for two dimensions}
\usage{
loading_plot(
  loadings,
  axis1 = 1,
  axis2 = 2,
  col_points = "blueviolet",
  constraints = NULL,
  arrows = TRUE,
  ...
)
}
\arguments{
\item{loadings}{The matrix of loadings to plot}

\item{axis1}{(default 1) column of \code{loadings} to plot on horizontal axis}

\item{axis2}{(default 2) column of \code{loadings} to plot on vertical axis}

\item{col_points}{(default "blueviolet") either a single color, or a vector of colors of length \code{NROW{loadings}}}

\item{constraints}{(default NULL) a list of constraints for the axes. See \code{\link[PTCA4CATA]{createFactorMap}}}

\item{arrows}{(default TRUE) Whether to draw arrows from the origin to the points}

\item{...}{args passed to \code{\link[PTCA4CATA]{createFactorMap}}}
}
\value{
A list containing:
  \item{\code{plot}}{The final plot created by the function}
  \item{\code{base_map}}{The output of \code{createFactorMap}}
  \item{\code{arrows}}{The output of \code{\link[PTCA4CATA]{addArrows}}}
}
\description{
Creates a factor map from \code{PTCA4CATA::createFactorMap} using \code{axis1} on the horizontal
 and \code{axis2} on the vertical.
}
\author{
Luke Moraglia
}
