% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsmfa.R
\name{mfa_norm}
\alias{mfa_norm}
\title{Normalize a matrix by its first singular value}
\usage{
mfa_norm(data, from = 1, to = NCOL(data))
}
\arguments{
\item{data}{The matrix to be normalized}

\item{from}{(default 1) If subsetting the columns of \code{data}, the starting index}

\item{to}{(default \code{NCOL(data)}) If subsetting the columns of \code{data}, the ending index}
}
\value{
A list containing
  \enumerate{
  \item{The normalized data}
  \item{The first singular value used to normalize}
  }
}
\description{
Divides a matrix by its first singular value, the norm used in MFA and PLSMFA.
}
\author{
Luke Moraglia
}
