% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_latent_variable.R
\name{partial_latent_variable}
\alias{partial_latent_variable}
\title{Compute partial latent variables for a subtable}
\usage{
partial_latent_variable(
  data_ori,
  loading_mat,
  from_index,
  to_index,
  n_subtable
)
}
\arguments{
\item{data_ori}{Should be the MFA normalized data, so either \code{mfa_norm_X} or \code{mfa_norm_Y}}

\item{loading_mat}{The corresponding matrix of loadings, either \code{u} or \code{v}}

\item{from_index}{The starting index of the subtable}

\item{to_index}{The ending index of the subtable}

\item{n_subtable}{The total number of subtables in \code{data_ori}}
}
\value{
The partial latent variables for the subtable
}
\description{
Whereas latent variables are computed as linear combinations of all the variables in a data
table, partial latent variables are computed from only the variables of a specific subtable.
}
\author{
Luke Moraglia
}
