% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsmfa.R
\name{plsmfa}
\alias{plsmfa}
\title{Perform partial least squares multiple factor analysis (PLSMFA)}
\usage{
plsmfa(
  data1,
  data2,
  column_design1,
  column_design2,
  center1 = TRUE,
  center2 = TRUE,
  scale1 = "SS1",
  scale2 = "SS1"
)
}
\arguments{
\item{data1}{The first data matrix, also called "X", with \eqn{N} rows and \eqn{I} columns
organized into \eqn{K} subtables.}

\item{data2}{The second data matrix, also called "Y", with \eqn{N} rows and \eqn{J} columns
organized into \eqn{M} subtables.}

\item{column_design1}{A vector of length \eqn{I} that gives the subtable membership of the X variables}

\item{column_design2}{A vector of length \eqn{J} that gives the subtable membership of the Y variables}

\item{center1}{(default TRUE) Whether to center \code{data1}}

\item{center2}{(default TRUE) Whether to center \code{data2}}

\item{scale1}{(default "SS1") Whether to scale \code{data1}.
Any acceptable input to \code{\link[ExPosition]{expo.scale}}}

\item{scale2}{(default "SS1") Whether to scale \code{data2}
Any acceptable input to \code{\link[ExPosition]{expo.scale}}}
}
\value{
A list containing:
  \item{\code{pls}}{results from the underlying PLSC analysis. This is output from \code{\link[GSVD]{gplssvd}}.
  It is a list containing:
   \describe{
    \item{\code{d}}{The singular values}
    \item{\code{u}}{The left singular vectors. For PLSMFA, these are the loadings of X.}
    \item{\code{v}}{The right singular vectors. For PLSMFA, these are the loadings of Y.}
    \item{\code{d_full}}{Should be the same as \code{d}}
    \item{\code{l_full}}{Should be the same as \code{l}}
    \item{\code{l}}{The eigenvalues}
    \item{\code{lx}}{The latent variables from X}
    \item{\code{ly}}{The latent variables from Y}
    \item{\code{p}}{The left generalized singular vectors. Should be the same as \code{u}}
    \item{\code{fi}}{The X variable "factor scores".
                     This is the left singular vectors multiplied by their singular values.}
    \item{\code{q}}{The right generalized singular vectors. Should be the same as \code{v}}
    \item{\code{fj}}{The Y variable "factor scores".
                     This is the right singular vectors multiplied by their singular values.}
    }
  }
  \item{\code{normed_X}}{X after MFA normalization of the subtables}
  \item{\code{normed_Y}}{Y after MFA normalization of the subtables}
  \item{\code{first_singval_X}}{A vector containing the first singular value of each subtable in X}
  \item{\code{first_singval_Y}}{A vector containing the first singular value of each subtable in Y}
  \item{\code{partial_lv_X}}{A 3D array, of size \eqn{N \times L \text{ (number of dimensions) } \times K}
   (subtables of X), containing the partial latent variables for the X subtables}
  \item{\code{partial_lv_Y}}{A 3D array, of size \eqn{N \times L \text{ (number of dimensions) } \times M}
   (subtables of Y), containing the partial latent variables for the Y subtables}
  \item{\code{design}}{A list containing information about the column designs including:
                \describe{
                \item{\code{column_design_X}}{the same as \code{column_design1} from input}
                \item{\code{column_design_Y}}{the same as \code{column_design2} from input}
                \item{\code{from_indices_X}}{a vector containing the column index that starts each subtable of X}
                \item{\code{to_indices_X}}{a vector containing the column index that ends each subtable of X}
                \item{\code{from_indices_Y}}{a vector containing the column index that starts each subtable of Y}
                \item{\code{to_indices_Y}}{a vector containing the column index that ends each subtable of Y}
                }
                }
   \item{\code{ctr}}{Contributions of variables and subtables. Output from \code{\link{contributions}}}
}
\description{
PLSMFA is a generalization of PLSC when the two data tables have subtables of variables.
The matrices \code{data1} and \code{data2} must have columns that are ordered by groups.
The column group memberships are stored in \code{column_design1} and \code{column_design2}.
If your data's columns are not ordered by their groups, but you have the out of order column
design vector, you can use \code{\link{arrange_data_by_design}} to order the columns.
}
\details{
PLSMFA implements PLSC with an additional preprocessing step that normalizes subtables within
the two data tables to be analyzed. The normalization is borrowed from multiple factor analysis (MFA),
which normalizes subtables by dividing them by their first singular value.

PLSMFA begins by centering
and scaling the variables as given in \code{center1}, \code{center2}, \code{scale1}, and \code{scale2}.
Then each subtable is decomposed by the SVD and divided by its first singular value (\code{\link{mfa_norm}}).
The cross-product matrix of the normalized data tables is computed and decomposed by the SVD.

The results are very similar to PLSC. However, \code{plsmfa} also computes the contribution of
each variable and each subtable. Additionally, each subtable has a corresponding matrix of partial latent
variables, which describe the observations as seen by only this subtable.
}
\author{
Luke Moraglia
}
